+++ changes.md
@@ -2,7 +2,7 @@
 
 ## Changes in dev
 1. `settings.OnlyLastCycle` defines if only last object version show in TFile (also as `lastcycle` URL parameter)
-2. Let change many settings parameters via "Settings" menu of the top hierarchy item
+2. Let change `settings` and `gStyle` parameters via "Settings" menu of the top hierarchy item
 3. Settings and gStyle can be stored as cookies, automatically read with every next load of website
 4. Support new `TGraph2DAsymmErrors` class
 5. Implement dark mode, can be activated with "&dark" parameters in URL or via settings menu